✨ feat: update pdf viewer
use viewer store
change layout
