+++ components/PdfViewer/OptionPanel/optionPanel.module.css
@@ -1,8 +1,8 @@
 .container {
   display: flex;
-  flex-direction: column;
+  flex-direction: row;
   align-items: center;
-  border-radius: 10px 10px 0 0;
+  border-radius: 0;
   width: 100%;
   background-color: var(--primary-dark);
   overflow: clip;
@@ -10,13 +10,21 @@
 }
 
 .handle {
-  cursor: row-resize;
-  height: 20px;
+  cursor: col-resize;
+  height: 100%;
+  width: 24px;
   flex: 0;
   display: flex;
   align-items: center;
   justify-content: center;
+}
+
+.content {
+  height: 100%;
   width: 100%;
+  flex: 1;
+  display: flex;
+  flex-direction: column;
 }
 
 .utils {
@@ -48,6 +56,7 @@
 }
 
 .keywords {
+  box-sizing: border-box;
   display: flex;
   justify-content: center;
   width: 100%;
@@ -63,10 +72,16 @@
   border-radius: 5px;
   background-color: var(--secondary-light);
   color: var(--primary-dark);
+  transition: background-color ease-out 0.3s;
+}
+
+.enabled {
+  background-color: var(--foreground);
+  color: var(--primary-dark);
 }
 
 .icon {
-  cursor: pointer;
   text-align: center;
   color: var(--background);
+  rotate: 90deg;
 }
